requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
mysql_connector_python==9.0.0
Pillow==10.4.0
Requests==2.32.3
Jinja2==3.1.4
